Moves name to be a part of BlenderDataBaseClass
The name property is part of the ID class in Blender (https://docs.blender.org/api/current/bpy.types.ID.html), which BlenderDataBaseClass at this point basically emulates.
